refactor: extract utils
