+++ src/keybindings/use-keybindings.ts
@@ -1,64 +1,7 @@
 import { useCallback, useEffect } from '@pionjs/pion';
 import { useMeta } from '../hooks/use-meta';
-import {
-	Activity,
-	ActivityHandler,
-	KeyBinding,
-	Matcher,
-	RegisterFn,
-} from './types';
-
-declare global {
-	interface ObjectConstructor {
-		entries(o: Matcher): [keyof Matcher, Matcher[keyof Matcher]][];
-	}
-}
-
-const matches =
-	(e: KeyboardEvent) =>
-	([matcher]: KeyBinding) =>
-		Object.entries(matcher).every(([key, value]) => e[key] === value);
-
-const isInteractive = (el: Element | null | undefined) => {
-	if (el == null) return false;
-
-	const bounds = el.getBoundingClientRect(),
-		root = el.getRootNode() as ShadowRoot | Document,
-		topEl = root.elementFromPoint(
-			bounds.x + bounds.width / 2,
-			bounds.y + bounds.height / 2,
-		);
-
-	return el === topEl;
-};
-
-const getActiveElement = (
-	root: Document | ShadowRoot = document,
-): Element | null => {
-	const activeEl = root.activeElement;
-
-	if (!activeEl) {
-		return null;
-	}
-
-	if (activeEl.shadowRoot) {
-		return getActiveElement(activeEl.shadowRoot);
-	}
-
-	return activeEl;
-};
-
-const focusIsInEditableArea = (): boolean => {
-	const active = getActiveElement(document);
-
-	if (!active) return false;
-	if (active.matches('input, textarea')) return true;
-	if ('isContentEditable' in active && active.isContentEditable) {
-		return true;
-	}
-
-	return false;
-};
+import { Activity, ActivityHandler, KeyBinding, RegisterFn } from './types';
+import { focusIsInEditableArea, isInteractive, matches } from './utils';
 
 type State = {
 	bindings: readonly KeyBinding[];
